requirements.txt
requests>=2.31.0
pandas>=2.2.0
python-dateutil>=2.9.0
tabulate>=0.9.0
